# backend/openapi.yaml
openapi: 3.0.1
info:
  title: AguaVigia CTG — API publica
  description: |-
    Monitoreo ciudadano del acueducto de Cartagena de Indias.

    Plataforma ciudadana e independiente: no esta afiliada a Aguas de Cartagena
    S.A. E.S.P. ni a ninguna entidad distrital.
  contact:
    name: Equipo AguaVigia CTG — Tecnologico Comfenalco
  license:
    name: Proyecto de aula
  version: 1.0.0
servers:
- url: /
  description: Mismo origen que sirve la API
...
paths:
...
  /api/veedor/segundo-factor/confirmacion:
    post:
      tags:
      - Veedor - Segundo factor
      summary: Confirmar el alta con un codigo de la app
      description: |-
        Devuelve una sesion nueva de alcance COMPLETO. Es lo que permite que un ADMIN
        recien sembrado pase de su sesion restringida al panel sin volver a escribir la
        clave que acaba de escribir.
      operationId: confirmar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudCodigo"
        required: true
      responses:
        "200":
          description: Segundo factor activo; sesion nueva emitida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "401":
          description: El codigo no coincide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
        "409":
          description: No hay un alta en curso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionVeedor"
  /api/veedor/segundo-factor/baja:
    post:
      tags:
      - Veedor - Segundo factor
      summary: Desactivar el segundo factor de la propia cuenta
      description: |-
        Exige un codigo valido: si bastara con la sesion, un token robado podria quitar
        de en medio justamente la defensa que impide usarlo. Un ADMIN no puede
        desactivarlo, su rol lo exige.
      operationId: desactivar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudCodigo"
        required: true
      responses:
        "204":
          description: Segundo factor desactivado
        "401":
          description: El codigo no coincide
        "409":
          description: El rol ADMIN exige segundo factor
...
  /api/reportes/{id}/confirmar:
    post:
      tags:
      - Reportes
      summary: Confirmar un reporte
      description: Permite a otro vecino confirmar un reporte ciudadano (M11).
      operationId: confirmar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudConfirmar"
        required: true
      responses:
        "200":
          description: Reporte confirmado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
        "400":
          description: Error en la solicitud
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
        "404":
          description: Reporte no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteRespuesta"
...
  /api/veedor/usuarios/{id}/aprobacion:
    patch:
      tags:
      - Veedor - Cuentas
      summary: "Aprobar una cuenta que ya verifico su correo, asignandole permisos"
      operationId: aprobar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitudPermisos"
        required: true
      responses:
        "200":
          description: Cuenta activa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioRespuesta"
        "409":
          description: "La cuenta no esta esperando aprobacion, o el ADMIN se administra\
            \ a si mismo"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioRespuesta"
  /api/veedor/reportes/{id}/descartar:
    patch:
      tags:
      - Veedor - Moderación
      summary: Descartar un reporte
      operationId: descartar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Reporte descartado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteModeracionRespuesta"
        "400":
          description: El reporte no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/veedor/reportes/{id}/aprobar:
    patch:
      tags:
      - Veedor - Moderación
      summary: Aprobar un reporte
      operationId: aprobar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Reporte aprobado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteModeracionRespuesta"
        "400":
          description: El reporte no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/veedor/ingesta/propuestas/{id}/descartar:
    patch:
      tags:
      - Veedor - Ingesta
      summary: Descartar una propuesta
      description: "No toca el sector. La propuesta se archiva como descartada, no\
        \ se borra."
      operationId: descartar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Propuesta descartada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropuestaIngestaRespuesta"
        "400":
          description: La propuesta no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/veedor/ingesta/propuestas/{id}/aprobar:
    patch:
      tags:
      - Veedor - Ingesta
      summary: Aprobar una propuesta
      description: |-
        Aplica el estado propuesto al sector y anexa el evento a la bitácora pública
        (RF026). Es el único camino por el que la ingesta llega al mapa.
      operationId: aprobar_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Propuesta aprobada y estado aplicado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropuestaIngestaRespuesta"
        "400":
          description: La propuesta no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: El sector de la propuesta ya no existe
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/veedor/reportes/pendientes:
    get:
      tags:
      - Veedor - Moderación
      summary: "Listar los reportes pendientes de moderación, más antiguos primero"
      description: |-
        Paginado, con el total y el enlace a la siguiente página en las cabeceras
        `X-Total-Count` y `Link`. Por defecto 50; el máximo por página es 200.
      operationId: listarPendientes
      parameters:
      - name: pagina
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: tamano
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReporteModeracionRespuesta"
...
  /api/veedor/ingesta/propuestas:
    get:
      tags:
      - Veedor - Ingesta
      summary: "Listar las propuestas pendientes de revisión, más recientes primero"
      description: |-
        Paginado, con el total y el enlace a la siguiente página en las cabeceras
        `X-Total-Count` y `Link`. Por defecto 50; el máximo por página es 200.
      operationId: listarPendientes_1
      parameters:
      - name: pagina
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: tamano
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Listado generado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PropuestaIngestaRespuesta"
...
  /api/suscripciones/confirmar:
    get:
      tags:
      - Suscripciones
      summary: Confirmar la suscripción (doble opt-in)
      description: |-
        Enlace del correo de confirmación. El token es de un solo enlace, no de un solo uso:
        confirmarla dos veces no falla (RF013). Responde JSON o una página HTML de cortesía
        según el `Accept` de quien pide — el mismo enlace del correo abre bien tanto en un
        navegador como desde un cliente de API.
      operationId: confirmar_2
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      - name: Accept
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Suscripción confirmada
          content:
            application/json:
              schema:
                type: object
            text/html:
              schema:
                type: object
        "400":
          description: "Token inválido, inexistente o de una suscripción ya cancelada"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    UsuarioRespuesta:
      type: object
      properties:
        id:
          type: string
        correo:
          type: string
        nombre:
          type: string
        estado:
          type: string
          description: "PENDIENTE_VERIFICACION, PENDIENTE_APROBACION, INVITADA, ACTIVA,\
            \ SUSPENDIDA o RECHAZADA"
        rol:
          type: string
        permisosEfectivos:
          type: array
          items:
            type: string
        permisosConcedidos:
          type: array
          items:
            type: string
        permisosRevocados:
          type: array
          items:
            type: string
        segundoFactorActivo:
          type: boolean
        creadoEn:
          type: string
          format: date-time
        actualizadoEn:
          type: string
          format: date-time
      description: "Cuenta del panel, tal como la ve un ADMIN"
...
    SesionVeedor:
      type: object
      properties:
        token:
          type: string
          description: "Token JWT. Se envia como 'Authorization: Bearer <token>'"
        usuarioId:
          type: string
        nombre:
          type: string
        correo:
          type: string
        rol:
          type: string
          description: "ADMIN, VEEDOR u OBSERVADOR"
        permisos:
          type: array
          description: "Permisos efectivos ya resueltos: rol mas concedidos menos\
            \ revocados"
          items:
            type: string
            description: "Permisos efectivos ya resueltos: rol mas concedidos menos\
              \ revocados"
        alcance:
          type: string
          description: |-
            COMPLETO, o ALTA_SEGUNDO_FACTOR cuando la cuenta es ADMIN y todavia no dio de alta su
            TOTP. Con ese alcance el token solo sirve para /api/veedor/segundo-factor.
      description: "Sesion emitida para el panel del veedor (RNF011: expira en 8 horas)"
    SolicitudCodigo:
      type: object
      properties:
        codigo:
          minLength: 1
          type: string
      description: Codigo de 6 digitos de la app de autenticacion
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
...
    CoordenadaDTO:
      required:
      - latitud
      - longitud
      type: object
      properties:
        latitud:
          maximum: 90
          exclusiveMaximum: false
          minimum: -90
          exclusiveMinimum: false
          type: number
          format: double
        longitud:
          maximum: 180
          exclusiveMaximum: false
          minimum: -180
          exclusiveMinimum: false
          type: number
          format: double
      description: "Coordenada GPS del reporte, solo cuando el usuario la autoriza\
        \ (RF007)"
...
    ReporteRespuesta:
      type: object
      properties:
        id:
          type: string
        sectorId:
          type: string
        tipo:
          type: string
        timestamp:
          type: string
          format: date-time
        fotoUrl:
          type: string
        confirmaciones:
          type: integer
          format: int32
      description: Reporte ciudadano registrado
    SolicitudConfirmar:
      type: object
      properties:
        huella:
          maxLength: 128
          minLength: 32
          type: string
          description: Huella hash del dispositivo del usuario que confirma (ADR-007)
          example: 3a7b9c...
      description: Solicitud para confirmar un reporte ciudadano por otro vecino
...
    SolicitudPermisos:
      type: object
      properties:
        rol:
          minLength: 1
          type: string
          description: "ADMIN, VEEDOR u OBSERVADOR"
        concedidos:
          type: array
          items:
            type: string
        revocados:
          type: array
          items:
            type: string
      description: |-
        Rol de base mas los ajustes por persona. Los permisos del rol se aplican solos; `concedidos`
        anade sobre ellos y `revocados` quita. Un permiso en las dos listas es un error y se rechaza.
    ReporteModeracionRespuesta:
      type: object
      properties:
        id:
          type: string
        sectorId:
          type: string
        tipo:
          type: string
        coordenada:
          $ref: "#/components/schemas/CoordenadaDTO"
        timestamp:
          type: string
          format: date-time
        estadoModeracion:
          type: string
          description: "PENDIENTE, APROBADO o DESCARTADO"
      description: Reporte ciudadano en la cola de moderación del veedor (RF018)
    PropuestaIngestaRespuesta:
      type: object
      properties:
        id:
          type: string
        sectorId:
          type: string
        estadoPropuesto:
          type: string
          description: "SIN_SERVICIO, PRESION_BAJA, CORTE_PROGRAMADO o CON_SERVICIO"
        fuente:
          type: string
          description: Colector que la detectó
          example: acuacar
        urlOriginal:
          type: string
          description: Enlace al boletín o nota de prensa original
          nullable: true
        citaTextual:
          type: string
          description: "Fragmento del que se dedujo el estado, para que el veedor\
            \ pueda verificarlo"
        confianza:
          type: number
          description: |-
            Entre 0 y 1, graduada según la evidencia que halló el extractor (ADR-032): 0.85 con
            enumeración explícita de barrios y horario, 0.75 con enumeración sin horario, 0.45
            con una mención suelta en prosa. Sirve para ordenar la cola, no para publicar solo.
          format: double
        detectadaEn:
          type: string
          format: date-time
        estadoRevision:
          type: string
          description: "PENDIENTE, APROBADA o DESCARTADA"
        inicioDeclarado:
          type: string
          description: |-
            Inicio de la ventana que el boletín prometió. Nulo cuando el texto no la declaraba:
            no se estima (ADR-006).
          format: date-time
          nullable: true
        finPrometido:
          type: string
          description: |-
            Fin prometido de la misma ventana. Junto con el inicio es lo que permite que el
            estado del sector evolucione solo (ADR-033) y lo que alimenta el Índice de
            Cumplimiento (RF020-RF022).
          format: date-time
          nullable: true
      description: |-
        Propuesta de cambio de estado detectada por la ingesta automatizada (M9), esperando la
        revisión de un veedor. No afecta el mapa público hasta que se apruebe.
...